kvm: split driver to put events into its own module
