+++ libvmi/driver/kvm/kvm.h
@@ -95,21 +95,6 @@ status_t kvm_pause_vm(
     vmi_instance_t vmi);
 status_t kvm_resume_vm(
     vmi_instance_t vmi);
-status_t kvm_events_listen(
-    vmi_instance_t vmi,
-    uint32_t timeout);
-status_t kvm_set_reg_access(
-    vmi_instance_t vmi,
-    reg_event_t* event);
-status_t kvm_set_intr_access(
-    vmi_instance_t vmi,
-    interrupt_event_t* event,
-    bool enabled);
-status_t kvm_set_mem_access(
-    vmi_instance_t vmi,
-    addr_t gpfn,
-    vmi_mem_access_t page_access_flag,
-    uint16_t vmm_pagetable_id);
 status_t kvm_get_vcpuregs(
     vmi_instance_t vmi,
     registers_t *regs,
@@ -150,10 +135,6 @@ driver_kvm_setup(vmi_instance_t vmi)
     driver.is_pv_ptr = &kvm_is_pv;
     driver.pause_vm_ptr = &kvm_pause_vm;
     driver.resume_vm_ptr = &kvm_resume_vm;
-    driver.events_listen_ptr = &kvm_events_listen;
-    driver.set_reg_access_ptr = &kvm_set_reg_access;
-    driver.set_intr_access_ptr = &kvm_set_intr_access;
-    driver.set_mem_access_ptr = &kvm_set_mem_access;
     vmi->driver = driver;
     return VMI_SUCCESS;
 }
